# swagger/jmarkets.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JMarkets API Documentation",
    "version": "1.0.0",
    "description": "Multi-tenant marketplace platform API with CMS, authentication, RBAC, and auto-deployment",
    "contact": {
      "name": "JMarkets Support",
      "url": "https://j-markets.jcampos.dev"
    }
  },
  "servers": [
    {
      "url": "https://markets-api.jcampos.dev",
      "description": "Production API"
    },
    {
      "url": "http://localhost:5000",
      "description": "Development API"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/api/users/{userId}/organization/{orgId}/deployments": {
      "post": {
        "summary": "Trigger auto-deployment for organization",
        "tags": [
          "Deployment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment triggered successfully"
          },
          "400": {
            "description": "Organization context required"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/{userId}/organization/{orgId}/deployments/status": {
      "get": {
        "summary": "Get deployment status for organization",
        "tags": [
          "Deployment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current deployment status"
          },
          "400": {
            "description": "Organization context required"
          }
        }
      }
    },
    "/api/users/{userId}/organization/{orgId}/deployments/history": {
      "get": {
        "summary": "Get deployment history for organization",
        "tags": [
          "Deployment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of deployments"
          },
          "400": {
            "description": "Organization context required"
          }
        }
      }
    },
...
  },
...
}